QemuCpu: fix issue with self-modifyig code in dmi regions
Previously, we were creating a ram_ptr qemu::MemoryRegion for each dmi region.
So each core ended up with its own qemu::MemoryRegion, actually pointing
to the same host ram.
This caused a bug when the guest was doing self-modifying code, because
QEMU did not know the different regions where actually pointing to the
same place, so TBs were not flushed accordingly.

This patch makes dmi regions to be created only once, and cores now create
aliases to this regions, so QEMU knows they are actually the same region.