port to new libqemu-cxx API
CPUs now provide a qemu::Target instead of a bare string to indicate the
QEMU target they are attached to. This commit removes the library loader
from libqbox as it is now provided by libqemu-cxx.

For now we keep a static instance of the loader. This static instance
will be removed once libqbox will be refactored and the loader made
explicit.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>