change target info cmake configuration
Using configure_file is not practical in our case. The target property
we read from libqemu-<target> targets may contain generator expression
(namely $<INSTALL_INTERFACE> and $<BUILD_INTERFACE>) that won't be
extended in the generated file. This is because of the way cmake works
(basically generator expressions are evaluated at build time, not at
cmake generation time).

Instead, use target_compile_definitions() to provide libraries path. It
requires a quite complex nested generator expression to:
  - retrieve the target property
  - expand the $<INSTALL_INTERFACE> and $<BUILD_INTERFACE> expressions
  - reduce the list to one element with $<JOIN>

The drawback of this method is having to add #ifndef...#define macros
into target_info.cc, but at the end, I prefer to have nullptr instead of
empty strings when a library is not supported.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>