Refactor the instance private data into a separate class
Add the LibQemuInternals class to hold the shared state between the
LibQemu instance and its objects. Before this commit, we were only
sharing the LibQemuExports struct, which holds all the function pointers
to the libqemu API.

For now, LibQemuInternals only have this LibQemuExports member, but
future commits will enhance it with other shared data.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>