start of refactoring: remove every source files
The refactoring is so big that it makes more sense to start from an
almost empty repository. So many things are wrong, misplaced, or should
not be here at all.

This commit removes every pieces of code, and rewrite the CMakeLists.txt
file to be modern CMake compliant and ready for the next step.

Next commits will re-add code bit by bit.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>