Add the QemuInstance and QemuInstanceManager classes
The QemuInstance class encapsulate a qemu::LibQemu instance and handles
argument passing to QEMU and instance initialization. This class is
meant to explicit the QEMU instance within QBox. Devices will be
contructed with an instance.

The QemuInstanceManager class helps in creating instances for a given
target. It handles the library loader so that loading multiple instances
of the same library is possible.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>