Cpu: add prepare_run_cpu and sync_with_kernel methods
Those two methods are the first step in constructing the CPU loop.

prepare_run_cpu is meant to be called before a loop run. It starts by
locking the iothread mutex, check if the CPU has work and release it by
setting soft_stopped to false. It finally rearm the deadline timer to
limit execution to the maximum run budget.

If the CPU cannot run (because it is e.g. waiting for interrupt), it
will go to sleep in wait_for_work. This method waits for an event on the
m_external_ev list. This list is meant to be filled with event related
to external activities that can wake the CPU up. Those events are CPU
specific so this list should be filled by children classes.

sync_with_kernel is meant to be called at the end of a loop run. It
deletes the deadline timer, stops the CPU, unlock the iothread mutex and
update the quantum keeper, synchronizing if needed.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>