Cpu: add the missing CPU loop pieces
Add the rest of the required pieces for the CPU to run. This implements
both coroutine and TCG thread mode.

In coroutine mode, we create a SystemC thread for the CPU loop
(mainloop_thread_coroutine). It calls prepare_run_cpu, run_cpu_loop and
sync_with_kernel in a loop. The end of loop callback is used to yield to
the main SystemC thread loop.

In TCG thread mode, we rely on the end of loop callback to call
sync_with_kernel and prepare_run_cpu. run_cpu_loop is not used in this
mode (QEMU has its own thread for the CPU and simply call our end of
loop callback so we have a chance to synchronize).

TCG thread mode needs extra care at start of simulation and at
destruction time. At start of simulation, we have to release the CPU so
it can run, it's similar to a call to prepare_run_cpu but we don't want
to mess with the iothread lock and wait_for_work. At destruction time,
we simply clear all the callbacks to be sure QEMU won't be calling us
after destruction (even if the underlying libqemu-cxx CPU object is
still alive). We also set unplug to true so that QEMU will eventually
exit the CPU thread. Unfortunately we cannot join the thread because of
some conditions leading to dead-locks (see comment in the destructor).

Signed-off-by: Luc Michel <luc.michel@greensocs.com>