Cpu: add the kick callback
The callback is called when the CPU is kicked in QEMU. It signals an
asynchronous event to wake the CPU up in case it is asleep in
wait_for_work. This allows to properly unblock the CPU thread when it is
waken up by e.g. the iothread (with an action from the GDB stub).

There is a race condition in this implementation as we release the
iothread mutex just before calling wait_for_work. In the mean time, the
kick could be triggered and the notification lost. Fixing this would
require to have another synchronisation mechanism as we cannot use the
classical "wait on condition" pattern here. The odds of this happening
are quite limited though as this is pretty much related to GDB stub
activity, which is controlled by human interaction most of the time.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>