Introduce the DMI manager
The QemuInstanceDmiManager class handles the DMI regions at the level of
a QEMU instance.

For a given DMI region, we need to use a unique memory region (called
the global memory region, in a sense that it is global to all the CPUs
in the instance). Each CPU is then supposed to create an alias to this
region to be able to access it. This is required to ensure QEMU sees
this region as a unique piece of memory.  Creating multiple regions
mapping to the same host address leads QEMU into thinking that those are
different data, and it won't properly invalidate corresponding TBs if
CPUs do SMC (self modifying code).

The LockedQemuInstanceDmiManager is the locked version of the previous
one to ensure safe concurrent accesses.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>