Introduce the QemuInitiatorSocket class
This class specializes the TLM-2.0 initiator socket to map a QEMU
AddressSpace to standard TLM-2.0 transactions.

For now, it only creates the root memory region that will be used to
intercept memory accesses from QEMU, and connects it to the device
AddressSpace designed by the property name given at init time.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>