QemuInitiatorSocket: add memory transaction support
This commits adds the wiring of QEMU to TLM-2.0 transaction. On a QEMU
I/O access, the corresponding callback is called. We unlock the iothread
mutex, create the TLM-2.0 transaction and forward it through the socket.
The actual transaction is done on the SystemC thread. When transaction
is done, we lock back the iothread mutex and do a final conversion of
TLM-2.0 to QEMU memory access return code.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>