QemuInitiatorSocket: add the QemuInitiatorIface interface
This interface allows for timing annotation during transactions. The two
methods initiator_get_local_time and initiator_set_local_time are used
respectively before and after the transaction to let the initiator
update its decoupled view of the time, and to correctly annotate the
b_transport call.

The initiator_customize_tlm_payload method gives a chance to the
initiator to customize the transaction before it is sent, e.g. to add
some extensions to it.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>