Introduce the QemuMrHintTlmExtension
This extension allows a target to signal that it is actually a QEMU
device and that the transaction accessed one of its memory region at a
given offset.

It is useful for a QEMU initiator so it can directly map the memory
region onto its root MR, improving general preformances as further
accesses won't go through the SystemC kernel.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>