QemuInitiatorSocket: add support for QemuMrHintTlmExtension
After a transaction, the socket checks for the presence of a
QemuMrHintTlmExtension in the TLM payload. If found, it directly maps
the returned memory region onto its root MR using an alias so that next
accesses won't go through the SystemC kernel.

Note that this optimization could be optionally disabled through a
parameter to allow easier debugging of memory transaction.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>