Introduce the QemuCpuHintTlmExtension
This extension is meant to be set by CPUs in transactions. Targets can
use this information to correctly set the current_cpu global variable
within the QEMU instance before performing the access.

It is required as some devices (like the ARM GICv2) relies on this
global variable to know which CPU if performing the access.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>