QemuCpu: add the default initiator socket
Every QEMU CPUs have a default address space attached to the "memory"
QOM property. This commit adds a socket attached to it. It also
implements the QemuInitiatorIface for the CPU as required by the
QemuInitiatorSocket.

Note that we have an issue in initiator_set_local_time as we cannot as
of now update the deadline timer after having updated the quantum
keeper. For some reasons it introduces performance penalties in icount
mode. For now, we just ignore the time passed into the transaction (so
we can in theory violate the time budget). This should be investigated
and fixed.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>