RunOnSysC: add sc_module_name parameter to constructor
SystemC enforces the presence of a sc_core::sc_module_name parameter to
get the hierarchy right. Moreover if we end up having multiple instance
of RunOnSysC in the same module, we must be able to name them
differently so that SystemC does not complain with name clashes. We
still provide a default value to avoid having to change all the users.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>