Introduce the QemuTargetSignalSocket class
Add the QemuTargetSignalSocket class that wraps a GPIO input on a QEMU
device. This class inherits from TargetSignalSocket<bool> so it can be
connected to classical sc_inout<bool> or sc_out<bool> ports, and to
TargetInitatorSocket<bool> sockets.

All modifications to this socket will be propagated to the wrapped QEMU
GPIO.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>