Introduce the QemuInitiatorSignalSocket class
Add the QemuInitiatorSignalSocket class that wraps a GPIO output on a QEMU
device. This class inherits from InitiatorSignalSocket<bool> so it can be
connected to classical sc_in<bool> port, and to
InitiatorInitatorSocket<bool> sockets.

All modifications to the wrapped QEMU GPIO is propagated through this
socket.

If this socket happens to be connected to a QemuTargetSignalSocket, the
propagation is done directly within QEMU and do not go through the
SystemC kernel. Note that this is only true if the GPIOs wrapped by both
this socket and the remote socket lie in the same QEMU instance.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>