cmake: disable submodule fetching for libqemu
By default, FetchContent gets submodule recursivelly. Moreover, CMake in
its current version (cmake version 3.19.7 at the time of writting) is
buggy and providing `GIT_SUBMODULES ""` with CMP0087 set to NEW does not
work as it should (not fetching any submodule).

The workaround to that is to provide an existing directory or file in
the git repository that is _not_ a submodule. At the end no submodule
are initialized, which is what we want. This is why this commits adds
this strange "GIT_SUBMODULE CMakeLists.txt" option to gs_addpackage.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>