QemuInitiatorSignalSocket: force event notification on direct connection
When the socket is connected to a QemuTargetSignalSocket, the GPIO event
is directly forwarded to QEMU without going through the SystemC kernel
(unless the initiator and the target are in different instances).

We still want the event to be signaled in case someone is waiting on it.
This commit adds a call to the notifiy method of the target socket, by
going through the SystemC kernel but without blocking.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>