Cpu: fix a race condition when going to sleep
There was a race condition in MTTCG mode when the CPU was going to sleep
by calling wait_for_work. If a wake-up event was triggered after the CPU
checked for work, but before the wait call on the SystemC kernel could
happen, the event was lost and the CPU thread would sleep forever.

This commit fixes this by introducing a m_signaled state to record the
fact that we have a pending event in MTTCG mode. A SystemC thread
watches the m_external_ev list and set this state when an event of this
list is signaled. The kick callback does the same.

Then, instead of using the SystemC kernel as a synchronisation point as
before, we now wait on m_signaled to become true when going to sleep.
Note that this is only true in MTTCG mode. The coroutine mode behaviour
hasn't changed.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>