DMI manager: rework DMI regions and aliases
In preparation for DMI invalidation, rework the way DMI regions and
aliases are handled.

This commit introduces two classes: DmiRegion and DmiRegionAlias. The
former represents a DMI region returned by a DMI request, before this
commit, it was named DmiInfo.

The DmiRegionAlias class represents an alias onto a DMI region. It
embeds the QEMU memory region an initiator must map onto its root MR to
setup the direct memory access.

The only method the QemuInstanceDmiManager now exposes is
get_new_region_alias to create a new alias based on a tlm::tlm_dmi
object.

The QemuInitiatorSocket is reworked to use this new API. More is now
done on the SystemC thread to ensure thread safety. Some parts (the MR
subregion mapping) must still be done on the QEMU CPU thread so the code
is written to accommodate this and avoid race conditions.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>