QemuCpu: fix a race condition where the kernel can starve
Fix a race condition where the kernel can starve while the CPU still has
work but did not wake up fast enough to restart its quantum keeper.

This commit fixes this by early restarting it in watch_external_ev().
This is not ideal but way simpler than waiting for the CPU thread to
wake up.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>