QemuCpu: fix a dead-lock when destructing the CPU
The CPU can dead-lock in the destructor waiting for the CPU thread to
finish is this one is stuck waiting for an I/O job to complete.

Cancel all I/O jobs in the destructor to unblock it.

Note: this is not ideal and should be reworked in case a CPU has
multiple sockets.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>