Introduce the testing infrastructure for libqbox
This commit adds the basic infrastructure for testing in libqbox. It
adds the TestBench class that future commits will use for testing.

A TestBench object is a top-level sc_module instanciated to run a test.
It is meant to be specialized by each test. It provides basic facilities
for testing. The convinient TEST_FAIL and TEST_ASSERT macro can be used
to either fail a test with a message or do a assertion that will fail
the test if evaluated to false.

The run_testbench function is meant to be called by sc_main, passing it
the command line arguments and instantiating its template parameter with
the test-bench class corresponding to the test you want to run. Thanks
to this function, the sc_main function becomes a one-liner to run a
test:

int sc_main(int argc, char *argv[])
{
    return run_testbench<MyTestBench>(argc, argv);
}

Signed-off-by: Luc Michel <luc.michel@greensocs.com>