test: add the CPU base test-bench infrastructure
This commit adds the CpuTestBench class. This test-bench instantiates
the following components:
  - The CPU under test
  - A router
  - A memory
  - A CPU tester

The CPU and the tester are specified using the two template arguments. A
CPU tester is a SystemC module meant to check one or more CPUs'
behaviour. It can have e.g. a target socket connected to to router so
the CPU can do memory accesses to it.

Next commits will add specialized testers for a given task, alongside
with the tests using them.

The test-bench allows to assemble a test firmware using the Keystone
library. The firmware is given as a string to the set_firmware method.
It is then assembled and loaded into memory.

Several CCI parameters allow configuring the test on the command line.
The corresponding qbox_add_cpu_test() CMake function will add multiple
occurrences of a test by varying those parameters (number of CPUs,
synchronization policy, icount mode on or off).

Signed-off-by: Luc Michel <luc.michel@greensocs.com>