QemuCpu: start in stopped mode when using the GDB stub
When starting a GDB stub, also stop the QEMU VM so that the execution
waits for GDB to connect. Also remove this check in end_of_loop_cb as it
would cause GDB's continue command to not work first time in some cases,
and is not useful anymore in the general case anyway.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>