Router: add support for the TLM exclusive extension
Make the router aware of the TLM exclusive extension. When detecting one
in a transaction, it calls add_hop on it with the ID corresponding to
the initiator from where the request comes from.

This effectively generates an unique global ID per initiator, whatever
the global address space layout is.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>