Add the TLM exclusive access extension
This extension allows to tag the current access as exclusive. An
exclusive access has an initiator ID and a exclusive store status.

The initiator ID is composed by the routers on the way of the
transaction, and the store status is filled by a target that supports
this extension to indicate whether the exclusive store succeeded or not.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>