QemuCpuArmCortexA53: add exclusive load/store support
This commit adds support to handle the exclusive load/store interaction
with the SystemC world.

Note that the ARM target QEMU implementation of the exclusive load/store
instruction is not totally accurate w.r.t the real hardware. Each CPU
keeps a copy of the memory at exclusive load time (env->exclusive_val)
along with the address of the exclusive access (env->exclusive_addr).
This value is compared to the current value of the memory at exclusive
store time. If the value hasn't changed, the exclusive store is
considered a success and the new value is written into memory. Otherwise
a failure is returned (this compare and exchange operation is done
atomically). No locking mechanism is involved from the point of view of
the global memory system.

On the SystemC side we can have this global lock mechanism though. We
have this exclusive access TLM extension we must use in case of an
exclusive access. The other end will indicate, in case of a store,
whether it succeeded or not.

This commits implement a trade-off to be functionally correct w.r.t
exclusive accesses on the SystemC side. First of all if the CPU does a
exclusive load/store pair on a DMI region, there is no way we can
guarantee anything. This is why if a region is locked on the SystemC
side, the corresponding DMI region must be invalidated (this is what the
GreenSocs ExclusiveMonitor component does).

If DMI is properly invalidated, we have a change to catch the exclusive
store access (and stores only!). This means that the CPU is still not
able to lock an exclusive region as we won't catch the exclusive loads.
This is not a problem since the local QEMU implementation of exclusive
accesses ensure functional correctness.

We can detect the exclusive stores using the following heuristic: since
QEMU is doing an atomic compare and exchange on exclusive store, we can
check that:
   1. The CPU is in an exclusive context.
   2. the env->exclusive_addr (set by the previous exclusive load)
      matches the current load/store address.

For 1., the way tcg_gen_atomic_cmpxchg_ixx translates depends on the
translation context. If QEMU had a memory pointer for the corresponding
region, it would translate it to some atomic host instruction. Since we
invalidated the DMI region for this store, the atomic access is not an
option and the corresponding translation block is re-translated in a
exclusive context (i.e. not a parallel one), leading to a regular
qemu_ld, followed by a qemu_st. This is why we check whether the CPU is
currently running in an exclusive context (meaning that other CPUs are
blocked). Note that this does _not_ work if QEMU is running in TCG
single thread mode as exclusive context is never entered in this case.
This has to be reworked for this case and is left as future work.

For 2., we simply ensure we are not intercepting another load/store.

Now that we know that we are dealing with a exclusive store, and we know
that we're going to have a load I/O, followed by a store I/O at the same
address, we can add the exclusive extension for both. The first one will
try to lock the exclusive region and the second one will report success
or failure (keep in mind we are already dealing with the exclusive
store!).

Finally, we report the failure to QEMU if needed. In this case, we have
to trick QEMU into thinking that the memory value has changed since the
exclusive load (even if it's not actually the case). We do so by
smashing the env->exclusive_val value so that it differs from the
current memory value. Due to some complications in the way QEMU
implements the checks env->exclusive_val after the atomic_cmpxchg, we
have to exit the TB so that it gets executed again with the modified
env->exclusive_val value.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>