QemuTargetSocket: use an AddressSpace to access the MemoryRegion
Accessing the memory region directly using
memory_region_dispatch_(read|write) is not enough. If a MR has some
sub-regions, those won't be resolved.

This commit adds an address space object and initializes it with the MR.
Then use the address space read/write function to access it. This way we
obtain the correct flat view of the memory.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>