QemuInitiatorSocket: do not assert when a DMI region already exists
We were previously expecting that if a DMI has been installed onto the
root MR, we would not have other I/O accesses on the same region. This
can be false though, e.g. when QEMU does a >8 bytes access. The access
is split into 8 bytes pieces. Even if the first one install a DMI
region, the rest will still go through I/O.

Simply remove the assert and add a log message so we can still debug
this if needed. Return nullptr so that the region is ignored.