Recovery and display of the parameters of the different parent and child brokers
-We add a CCI type to handle a ConfigurableBroker *
-We add a pointer to a cci param of a ConfigurableBroker * into the
ConfigurbaleBroker
-We look to see if we're a 'child' broker in the constructor, and if we are, we
'new' a cci_param<ConfigurableBroker *> - with the value "this"
The cci param will be 'registered' with the 'parent' (as we have not yet
registered this broker).
Then in the print_help() function, we look to see if the param has the name
childbroker and then we 'recurse' down into the child to print the param.

Signed-off-by: Thomas Marceron <thomas.marceron@greensocs.com>