libqbox/components: Add QEMU OpenCores Eth MAC
This exports the OpenCores Eth Model from Qemu.
It has to be used with the global peripheral initiator
socket, because Eth is a bus master that accesses
system RAM for DMA. If the transactions are not routed
through to SystemC they go to the wrong memory region
in Qemu and the device will not work.

Signed-off-by: Lukas Jünger <lukas.junger@greensocs.com>