WIP add qtimer tests
Right now this is still broken, because the cfgTable and cfgExtension
in hexagon_testboard.c are not initialized. Sadly, hexagon uses
callbacks that use these structs to clear interrupts. Thus this
segfaults. However, if you hardcode the initialization of the cfgTable
and cfgExtenion pointer to the values we use in main.cc, the test
somewhat works. At least the second hextimer of the qtimer signal
interrutps. To fix this we could pass down the structs into qemu or
rewrite this mechanism completely

Signed-off-by: Lukas Jünger <lukas.junger@greensocs.com>