Fix timer build on GCC 9.4:
For some reason GCC 9.4 does not understand the RAII syntax
used to initialize the arrays and wants to call the copy constructor.
Unfortunately, sc_export cannot be copied because the copy constructor
is private. I could not figure out how to make GCC understand that
this is an inplace construction and not a copy. Finally, I replaced
the arrays with multiple instances. It's not pretty, but it works.

Signed-off-by: Lukas Jünger <lukas.junger@greensocs.com>