Add basic interaction test.
This commit enables interaction tests to be run on GitLab CI
as well as locally. Currently only fastrpc_calc_test is
enabled and this is only tested against Makena configuration.

The commit allows the commit id of the firmware-repo used for
interaction tests to be pinned & tracked via a file present at
tests/firmware-repo-sha.txt

It is assumed that the tests are run in a docker container
that has a valid ~/.ssh directory and ~/.ssh/known_hosts file
with keys of gitlab.qualcomm.com & artifactory-las.qualcomm.com
and git-lfs installed.
Corresponding changes are made in the containers repository.

- QQVP_FIRMWARE_REPO_SHA : this specifies the git commit id
  in firmware-images repo. All tests are run against the
  version of binaries present in the given commit id.
  This CI variable can be set at the project/group level and
  can also be overridden per pipeline for manual runs.
  Note: the commit id specified by this variable overrides
  the one specified in the tests/firmware-repo-sha.txt file.

- LFS_SSH_PRIVATE_KEY : this must be set via Project/Group
  settings in GitLab UI. Since the key contents have special
  characters and newlines, they cannot be marked as "masked".
  GitLab documentatoin also doesn't address masking.
  The corresponding public key needs to be registered with
  GitLab as well as Artifactory-LAS for the service account
  we want to use - currently this is 'qqvpcibot' which is a
  new service account created just for this purpose.

- QQVP_FIRMWARE_REPO_DIR : This must be set to the base of the
  firmware-images repo. The tests reference the subdirectories
  within this path. Note that this must be set *before* cmake
  build step. Otherwise, the tests are disabled and `make test`
  will pass trivially.
  In CI this is set by .clone_firmware_images_repo job template.
  This env var can be used for local development as well.