Normalize reference configs
The reference config files have ram_* segments that are required
by the vBSP.  And now we've updated the reference configs to
match the tested configuration.  The no-GPU case is also included
as a reference for those who want to develop in a system without
a host GPU or without X available.

Signed-off-by: Brian Cain <bcain@quicinc.com>