Add a stress test
This test case is for QNX guests like our Makena or Lemans images.

You should use it by first doing a `make install`, checking out a copy
of the firmware-images repo, and then point Stress_test.py to the QQVP
install-under-test.

For example:

    ./tests/Stress_test.py -p lemans -f ../firmware-images/ -q ./vp_install_qemu/

Because of QTOOL-91621, you may need to copy librpc.so from your build to
$CMAKE_INSTALL_PREFIX/lib until we address that issue.

Once we are able to pass this test more reliably we can enable it as a
regression test.

Signed-off-by: Brian Cain <bcain@quicinc.com>