Extend interaction test timeout from 120s to 180s
Recent tests of quic-vp CI have encountered timeouts.  Previous runs
were marginal - some configurations would run for over 110s of the
alotted 120s time.

The Interaction_test.py timeout value is more closely bound to the
use case and the most appropriate place to set the timeout to a
"reasonable" value while allowing for varying CI host load.

The tests/CMakeLists.txt timeout value is a bit more distant from the
test scope.  The cmake value should only be intended to cause CI to
recover in case of a test design error (and not look for QQVP defects).
As such this timeout can be much higher.