Merge contiguous DMI regions
When a new DMI region is requested, we check whether it is contiguous
with other regions created previosly. In this case, we proceed with the
merge process.

Merge means we "delete" the old DMI region, and create a new DMI region
which includes both the old and the new region requested. Deletion of a
memory region does not happen immediately as it might still be in use.
This means the new region is going to "overlap" the old region for a
period of time, which requires us to use the right QEMU interface.

QEMU provides `memory_region_add_subregion_overlap()` which allows
adding overlapping regions to a container with a priority mechanism.
Indeed we introduced a root container in the DMI manager which is going
to contain all DMI regions. A new memory region which overlaps an old
one, will have a higher priority, and it will be preferred to the old
one while it is still alive.

Finally, all DMI aliases created for the initiator will alias the root
container, breaking the one-to-one mapping between DMI aliases and DMI
regions.

Signed-off-by: Antonio Caggiano <quic_acaggian@quicinc.com>
Signed-off-by: Mark Burton <mburton@quicinc.com>