[smmu] correct the SMR count
During testing of new configurations it was discovered that the
guest software for the QNX configuration had been patched to avoid
using too many CBs on QQVP.  But this was only because the QQVP
SMMU was misconfigured.

SMR count is based on register banks like "SYS_TCU_CFG_V2APPS_SMMU_SMRn",
which is shown as having 224 entries in ipcat.

CB count is based on register bannks like "SYS_TCU_CFG_V2APPS_SMMU_CBARn",
which is shown as having 128 entries in ipcat.

Signed-off-by: Brian Cain <bcain@quicinc.com>