Display module
Introduce a new SystemC module for abstracting the display from a
window system management point of view.

MacOS requires interacting with the window system exclusively from the
"main thread". This module runs these operations on SystemC kernel
thread while being able to interact properly with a QEMU instance
requiring a display backend.

A test is provided to verify that, while QEMU runs in its
own iothread, the window is correctly created on the main thread.

Signed-off-by: Antonio Caggiano <quic_acaggian@quicinc.com>