Switch priority to QUP UART
Previously the PL011 UART took precedence if both were enabled by
the configuration.  Now, the QUP UART takes precedence.  This means
that stdin is delivered to the QUP UART instead of the PL011.  So
older images which put a shell on the PL011 UART won't interact anymore
after this change, unless the QUP UART is disabled.

Signed-off-by: Brian Cain <bcain@quicinc.com>