Move the forwarded debug agent host ports lower
In the existing SA8775P configuration, host TCP ports 65535 and 65534 are
forwarded into the guest for the sake of the CDSP debug agent.  This appears
to conflict with a service already listening on port 65535.

    $ sudo lsof -i -P +c 25 | grep LISTEN | grep :65535
    ESDaemonBundle              545           root    3u  IPv4 0xf312c05ac14f33f      0t0    TCP *:65535 (LISTEN)

Before this change, vp would fail on macOS with:

    Could not set up host forwarding rule 'tcp::65535-:65535

Note: this change was made for a macOS system, but this is a general risk for
any host.  Different systems may have listeners installed on nearly any port,
so if there's an intersection, users will need to choose another port.  Also
- if there's another instance of vp running with this configuration, these
ports will already be bound to a listener and it will fail the same way.

Add more content to the user guide about how to address issues when port
conflicts come up.

Signed-off-by: Brian Cain <bcain@quicinc.com>