MemoryListener
Introduce the MemoryListener class which directly maps to the QEMU
MemoryListener. Note that this object is not a QEMU Object so it must be
created using the memory_listener_new method on the LibQemu class.

At the moment the only callback which can be set is for listening to
address space map events.

Signed-off-by: Antonio Caggiano <quic_acaggian@quicinc.com>