Remove libqemu from package-lock
CPM pulls libqemu from Qbox's CmakeLists.txt which expects a certain git
tag, but this tag is overwritten by the QQVP's package lock. This means
that when QQVP reference an older libqemu tag, CPM pulls this one and
Qbox would fail to compile or it would not work as expected.

Since there is no need to directly reference libqemu in QQVP's
package-lock, we can safely remove it and rely on Qbox's package-lock.

Signed-off-by: Antonio Caggiano <quic_acaggian@quicinc.com>