[SDPT-69] don't "fail" on --help and enforce --gs_luafile as required
Currently, running `vp --help` with no other arguments displays the
following error after the help message:

  Error: Loader: Unable to get parameter platform.hexagon_cluster_0.l2vic.mem.address
  In file: _deps/libgsutils-src/include/greensocs/gsutils/cciutils.h:157

This is because vp will try to load the luafile to report on the
configurable parameters, but no luafile was specified. There are two
issues here: 1) users should be able to run --help to understand the vp
usage without any extra args; 2) the error message for a missing
--gs_luafile argument is not too helpful.

To fix 1), let's introduce a new "--debug" option which will be
responsible for reporting on the state of the configurable parameters at
the beginning of the simulation and then halt the vp; while making
"--help" responsible only for printing the help message itself. This
allows "--help" to be used without any extra arguments.

And to fix 2), let's make clear that "--gs_luafile" is a required
argument for any vp execution besides "vp --help", printing an
appropriate error message when it is missing. Note that this fix comes
in two parts: this patch at qbox, and an additional change at qqvp to
instantiate the ConfigurableBroker with enforce_config_file=true. We
don't set this parameter to true unconditionaly here because some tests
actually do not need to enforce --gs_luafile.

Note: 'std::cout << std::endl' is practically the same as
'std::cout << '\n' << std::flush' (see [1]). So we clean up the help
message a bit by avoiding std::endl and the many 'std::cout <<' with
string concatenation.

[1]: https://cplusplus.com/reference/ostream/endl/

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>