Initialize m_val field
The "m_val" field was previously uninitialized but could be used-uninitialized
in the write() method.  When running with UBSan enabled, it would show the
runtime error indicated below:

    Info: SystemC: MASTER WRITE - Addr(SPI_TX_TRANS_LEN) and value:26c  1

    Info: SystemC: MASTER Addr(GENI_M_CMD_0):8000000
    /local/mnt/workspace/vp/qbox/include/greensocs/gsutils/ports/target-signal-socket.h:152:32: runtime error: load of value 190, which is not a valid value for type 'bool'

    Info: SystemC: MASTER WRITE - Addr and value:700  12

This error was detected because the uninitialized memory happened to contain
a value that's not legal for 'bool' and could have resulted in dropped signals.

While we're here, initialize the other members of this class in the
constructors.

Signed-off-by: Brian Cain <bcain@quicinc.com>