Use memcpy() to copy PL011 values
On some architectures, unaligned loads and stores are not permitted.  Copying
values directly by dereferencing pointers will result in loads and stores
to whatever address is passed.  The interface to the uart does not guarantee
aligned host addresses.

This was detected by UBSan and resulted in the error message below:

    /builds/qqvp/Qbox/include/greensocs/systemc-uarts/uart-pl011.h:169:28: runtime error: load of misaligned address 0x7fd95000b8cb for type 'uint32_t', which requires 4 byte alignment

This fix should improve the portability to other host architectures than
x86_64.

Signed-off-by: Brian Cain <bcain@quicinc.com>