Fix pci-tool check in Interaction_test.py and add rtl8139 PCI device check
The QNX /mnt/bin/pci-tool needs some environment variables to be set in order to
work successfully. A login non-interactive shell can be used from ssh command to
source /etc/profile and use the $(env) environment variables which exists in a normal
non-ssh interactive session. So, the command used should be:
$ ssh -p 2222 root@localhost "sh -l -c '/mnt/bin/pci-tool -v'"
Also the single quotes enclosing the pci-tool is important for the sh to execute
the command correctly, if omitted, then the pci-tool command will be executed without '-v'.
A check was added for the rtl8139 PCI device.